ticket:1275 - refactor gdal logic to the conversion module
Summary:
This patch does some refactoring in conversion/qlparser:
- the conversion code based on GDAL in QtEncode/QtDecode has been moved to conversion/gdal.cc
- QtEncode/QtDecode are removed altogether, everything is handled by QtConversion
- utility conversion classes in qlparser/ are moved to conversion/
- added --disable-gdal so that compiling rasdaman without GDAL is possible if needed
- fix hdf converter; remove duplicate files; remove tor and vff converters; move nitf code to separate directory

Test Plan: the existing conversion/select tests pass

Reviewers: vmerticariu, mdumitru

Reviewed By: mdumitru

Differential Revision: http://codereview.rasdaman.org/D187

+++ qlparser/qtnode.hh
@@ -20,7 +20,7 @@
 
 // define used in lots of qlparser files to indent output
 #ifdef CPPSTDLIB
-#define SPACE_STR(numSpace) std::string(numSpace,' ')
+#define SPACE_STR(numSpace) std::string((size_t)numSpace,' ')
 #else
 #define SPACE_STR(numSpace) std::string(' ',numSpace)
 #endif
